Update README.md
We would like to promote the IR21 via CU Denver's ESR by creating a trackable link and trackable QR code. CU Denver will be using Odder as a platform for this year's specific campus ESR, so there may be other digital opportunities with this new platform that Alex and Ian could be involved in.